Break out bulk edit tests and make them more robust
